requirements.txt
# filepath: /Users/suya/Desktop/Baduanjin_711/requirements.txt
flask
werkzeug
torch
torchvision
Pillow
google-generativeai
langchain-google-genai
gunicorn